+++ button.lua
@@ -10,16 +10,15 @@ function PauseGame()
         Buttons = {}
         State = 'pause'
         Buttons['resume'] = Button(10,50,200,50,"Resume",ResumeGame,0)
-        Buttons['options'] = Button(10,120,200,50,"Options",OptionsMenu,0.15)
-        Buttons['quit'] = Button(10,190,200,50,"Quit",SureQuit,0.3)
+        Buttons['options'] = Button(10,120,200,50,"Options",OptionsMenu,0.1)
+        Buttons['quit'] = Button(10,190,200,50,"Quit",SureQuit,0.2)
     end
 end
 
 function ResumeGame()
     if State == 'pause' then
         Buttons = {}
-        State = 'resuming'
-        ResumeTimer = SecondsCounter + 0.25
+        State = 'game'
 
     end
 end
@@ -29,7 +28,7 @@ function SureQuit()
         Buttons = {}
         State = 'surequit'
         Buttons['back'] = Button(10, 50, 200, 50, "Back", PauseGame, 0)
-        Buttons['quit'] = Button(10, 120, 200, 50, "Quit", love.event.quit, 0.2)
+        Buttons['quit'] = Button(10, 120, 200, 50, "Quit", love.event.quit, 0.1)
     end
 end
 
@@ -38,11 +37,11 @@ function OptionsMenu()
         Buttons = {}
         State = 'options' 
         Buttons['fullscreen'] = Button(10, 50, 300, 50, function() local x = 'Off' if love.window.getFullscreen() then x = 'On' end return "Fullscreen: "..x end, function() love.window.setFullscreen(not love.window.getFullscreen()) end, 0)
-        Buttons['vsync'] = Button(10, 120, 300, 50, function() local x = 'Off' if love.window.getVSync()==1 then x = 'On' end return "Vsync: "..x end, function() love.window.setVSync(1 - love.window.getVSync()) end, 0.1)
-        Buttons['renderer'] = Button(10, 190, 300, 50, function() local x = 'Screen' if NewRenderer then x = 'Update' end return "Renderer: "..x end, function() NewRenderer = not NewRenderer end, 0.2)
-        Buttons['graphics'] = Button(10, 260, 300, 50, function() local x = 'Fast' if HighGraphics then x = 'Fancy' end return "Graphics: "..x end, function() HighGraphics = not HighGraphics end, 0.3)
-        Buttons['creative'] = Button(10, 330, 300, 50, function() local x = 'Off' if CreativeMode then x = 'On' end return "Creative: "..x end, function() CreativeMode = not CreativeMode end, 0.4)
-        Buttons['back'] = Button(10, 400, 300, 50, "Back", PauseGame, 0.5)
+        Buttons['vsync'] = Button(10, 120, 300, 50, function() local x = 'Off' if love.window.getVSync()==1 then x = 'On' end return "Vsync: "..x end, function() love.window.setVSync(1 - love.window.getVSync()) end, 0.05)
+        Buttons['renderer'] = Button(10, 190, 300, 50, function() local x = 'Screen' if NewRenderer then x = 'Update' end return "Renderer: "..x end, function() NewRenderer = not NewRenderer end, 0.1)
+        Buttons['graphics'] = Button(10, 260, 300, 50, function() local x = 'Fast' if HighGraphics then x = 'Fancy' end return "Graphics: "..x end, function() HighGraphics = not HighGraphics end, 0.15)
+        Buttons['creative'] = Button(10, 330, 300, 50, function() local x = 'Off' if CreativeMode then x = 'On' end return "Creative: "..x end, function() CreativeMode = not CreativeMode end, 0.2)
+        Buttons['back'] = Button(10, 400, 300, 50, "Back", PauseGame, 0.25)
     end
 end
 
@@ -62,29 +61,29 @@ function Button:new(x, y, width, height, text, action, delay)
     self.text = text
     self.action = action
     self.hover = false
-    self.x = 0
-    self.y = 0
-    self.width = 0
-    self.height = 0
+    self.xpos = -5000
+    self.ypos = 0
+    self.width = width*GameScale
+    self.height = height*GameScale
 
     self.timeAlive = -delay or 0
 end
 
 function Button:update(dt)
     self.timeAlive = self.timeAlive + dt
-    self.x = self.xT * GameScale
 
     --slide in from left
-    
-    local slideInDistance = 16 * self.width * GameScale
-    if self.timeAlive < 0.5 then
-        self.x = self.x - slideInDistance * (0.5 - math.max(0,self.timeAlive))^3 -- slide in effect
+    local slideInDistance = 42 * self.width * GameScale
+    if self.timeAlive < 0.3 then
+        self.xpos = self.xT - slideInDistance * (0.3 - math.max(0,self.timeAlive))^3 -- slide in effect
+    else
+        self.xpos = self.xT
     end
 
-    self.y = self.yT * GameScale
+    self.ypos = self.yT * GameScale
     self.width = self.widthT * GameScale
     self.height = self.heightT * GameScale
-    if love.mouse.getX() > self.x and love.mouse.getX() < self.x + self.width and love.mouse.getY() > self.y and love.mouse.getY() < self.y + self.height then
+    if love.mouse.getX() > self.xpos and love.mouse.getX() < self.xpos + self.width and love.mouse.getY() > self.ypos and love.mouse.getY() < self.ypos + self.height then
         self.hover = true
         if love.mouse.isDown(1) then
             self:click()
@@ -101,13 +100,13 @@ function Button:draw()
         love.graphics.setColor(1,1,1,0.5)
     end
 
-    love.graphics.rectangle("fill",self.x,self.y,self.width,self.height,10,10)
+    love.graphics.rectangle("fill",self.xpos,self.ypos,self.width,self.height,10,10)
     love.graphics.setColor(0,0,0,1)
-    love.graphics.rectangle("line",self.x,self.y,self.width,self.height,10,10)
+    love.graphics.rectangle("line",self.xpos,self.ypos,self.width,self.height,10,10)
     if type(self.text) == 'string' then
-        simpleText(self.text,22,self.x + self.width/2,self.y + self.height/2,'center')
+        simpleText(self.text,22,self.xpos + self.width/2,self.ypos + self.height/2,'center')
     else
-        simpleText(self.text(),22,self.x + self.width/2,self.y + self.height/2,'center')
+        simpleText(self.text(),22,self.xpos + self.width/2,self.ypos + self.height/2,'center')
     end
     love.graphics.setColor(1,1,1,1)
 end